+++ x/inflation/keeper/sudo.go
@@ -5,6 +5,7 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
+	epochstypes "github.com/NibiruChain/nibiru/x/epochs/types"
 	inflationtypes "github.com/NibiruChain/nibiru/x/inflation/types"
 )
 
@@ -57,6 +58,18 @@ func (k sudoExtension) ToggleInflation(
 	}
 
 	params.InflationEnabled = enabled
+	if k.NumSkippedEpochs.Peek(ctx) == 0 {
+		// update skipped epochs since this means that inflation could have
+		// added to the chain after launch
+
+		epochInfo, err := k.epochsKeeper.GetEpochInfo(ctx, epochstypes.DayEpochID)
+		if err != nil {
+			return err
+		}
+
+		k.NumSkippedEpochs.Set(ctx, epochInfo.CurrentEpoch)
+	}
+
 	k.Params.Set(ctx, params)
 	return
 }